Refactor: 공통 기능 클래스 common 모듈로 이동 및 collection-system 패키지 경로 수정
